test fonts and storage in chromatic
