+++ .storybook/preview.js
@@ -1,49 +1,21 @@
 /* eslint-disable import/prefer-default-export */
 
-import React from 'react';
 import { addDecorator } from '@storybook/react';
 import { create } from '@storybook/theming';
-import * as fontFaces from '@bbc/psammead-styles/fonts';
-import GlobalStyles from '@bbc/psammead-styles/global-styles';
-import Cookies from 'js-cookie';
+import isChromatic from 'chromatic/isChromatic';
 
-const fontPathMap = [
-  { prefix: 'F_ISKOOLA_POTA_BBC', path: 'fonts/IskoolaPota/' },
-  { prefix: 'F_LATHA', path: 'fonts/Latha/' },
-  { prefix: 'F_MALLANNA', path: 'fonts/Mallanna/' },
-  { prefix: 'F_NOTO_SANS_ETHIOPIC', path: 'fonts/NotoSansEthiopic/' },
-  { prefix: 'F_PADAUK', path: 'fonts/Padauk/' },
-  { prefix: 'F_REITH_QALAM', path: 'fonts/ReithQalam/' },
-  { prefix: 'F_REITH_SANS', path: 'fonts/Reith/' },
-  { prefix: 'F_REITH_SERIF', path: 'fonts/Reith/' },
-  { prefix: 'F_SHONAR_BANGLA', path: 'fonts/ShonarBangla/' },
-];
+import clearBrowserStorage from './helpers/clearBrowserStorage';
+import loadFonts from './helpers/loadFonts';
 
-const clearBrowserStorage = () => {
-  Cookies.remove('ckns_policy');
-  Cookies.remove('ckns_explicit');
-  Cookies.remove('ckns_privacy');
-  window.localStorage.removeItem(`amp-store:${window.location.origin}`);
-};
+loadFonts();
+clearBrowserStorage();
 
 addDecorator(story => {
-  clearBrowserStorage();
+  if (isChromatic()) {
+    clearBrowserStorage();
+  }
 
-  return (
-    /* eslint-disable react/jsx-filename-extension */
-    <>
-      <GlobalStyles
-        fonts={Object.values(fontFaces).map(fontFace => {
-          const fontMap =
-            fontPathMap.find(map => fontFace.name.startsWith(map.prefix)) ||
-            fontPathMap[0];
-          return fontFace(fontMap.path);
-        })}
-      />
-      {story()}
-    </>
-    /* eslint-enable react/jsx-filename-extension */
-  );
+  return story();
 });
 
 const theme = create({
