+++ package.json
@@ -26,7 +26,7 @@
     "build:profile": "rm -rf build && cp envConfig/local.env .env && IS_PROD_PROFILE=true NODE_ENV=production webpack",
     "build:live": "cp envConfig/live.env .env && NODE_ENV=production webpack",
     "build:live:debug": "rm -rf build && awk '{sub(/LOG_DIR=.+/,\"LOG_DIR='log'\")}1' envConfig/live.env > .env && NODE_ENV=production webpack",
-    "build:storybook": "node .storybook/buildFontPreloads && build-storybook -s .storybook/static -c .storybook -o storybook_dist",
+    "build:storybook": "node .storybook/helpers/buildPreviewHead && build-storybook -s .storybook/static -c .storybook -o storybook_dist",
     "build:test": "cp envConfig/test.env .env && NODE_ENV=production webpack",
     "build:test:debug": "rm -rf build && awk '{sub(/LOG_DIR=.+/,\"LOG_DIR='log'\")}1' envConfig/test.env > .env && NODE_ENV=production webpack",
     "ci": "rm -rf node_modules && yarn install --frozen-lockfile",
@@ -41,7 +41,7 @@
     "preinstall": "node scripts/check-package-manager.js",
     "start": "NODE_ENV=production node build/server.js",
     "stop": "./scripts/killApp.sh",
-    "storybook": "node .storybook/buildFontPreloads && start-storybook -p 9001 -s .storybook/static -c .storybook",
+    "storybook": "node .storybook/helpers/buildPreviewHead && start-storybook -p 9001 -s .storybook/static -c .storybook",
     "test": "yarn build && yarn test:local",
     "test:local": "yarn test:lint && yarn test:dependencies && yarn test:unit && run-p --race start amp:validate && run-p --race start test:integration:ci",
     "test:dependencies": "node ./scripts/dependencyCheck",
@@ -137,6 +137,7 @@
     "moment": "2.29.1",
     "moment-timezone": "0.5.33",
     "morgan": "1.10.0",
+    "mustache": "^4.2.0",
     "path-to-regexp": "6.2.0",
     "polyfill-crypto.getrandomvalues": "1.0.0",
     "ramda": "0.27.1",
@@ -195,6 +196,7 @@
     "cypress-multi-reporters": "1.5.0",
     "cypress-terminal-report": "3.1.1",
     "depcheck": "1.4.1",
+    "ejs": "^3.1.6",
     "eslint": "7.26.0",
     "eslint-config-airbnb": "18.2.1",
     "eslint-config-prettier": "8.3.0",
